+++ NEWS.md
@@ -24,5 +24,5 @@ v4 (2012-10-19)
 v5 (????-??-??)
 - code: investigate if the inner-border can be retrieved
 - bugfix: [terminal on wrong monitor (github #10)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/10)
-- bugfix: animate the height instead of the position if there is a monitor above the primary one 
-
+- bugfix: better animation depending on monitor position (animate the height instead of the position if there is a monitor above the primary one)
+- bugfix: glitch at animation startup (the window can be briefly seen at the wrong location)